LayoutTests/media/media-source/media-source-minimumupcomingpresentationtime-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(updateend)
RUN(internals.setMaximumQueueDepthForTrackID(sourceBuffer, 1, 3))
RUN(sourceBuffer.appendBuffer(samples))
EVENT(updateend)
EXPECTED (internals.minimumUpcomingPresentationTimeForTrackID(sourceBuffer, 1) == '1') OK
RUN(video.currentTime = 4)
EVENT(seeked)
EXPECTED (internals.minimumUpcomingPresentationTimeForTrackID(sourceBuffer, 1) == '8') OK
RUN(sourceBuffer.appendBuffer(samples))
EVENT(updateend)
EXPECTED (internals.minimumUpcomingPresentationTimeForTrackID(sourceBuffer, 1) == '7') OK
END OF TEST
